chore(test): modify auth_test.
