--- internal/handlers/sms_test.go
@@ -1,1 +1,0 @@
-package handlers
