+++ internal/middleware/auth_test.go
@@ -20,16 +20,16 @@ func generateTestToken(email string, secret []byte, expired bool) string {
 
 	claims := &handlers.Claims{
 		Email: email,
-		Sub: email,
-		Name: "test user",
-		Iss: "customer-order-api",
-		Aud: "customer-order-api",
-		Exp: expirationTime.Unix(),	
-		Iat: time.Now().Unix(),
+		Sub:   email,
+		Name:  "test user",
+		Iss:   "customer-order-api",
+		Aud:   "customer-order-api",
+		Exp:   expirationTime.Unix(),
+		Iat:   time.Now().Unix(),
 		RegisteredClaims: jwt.RegisteredClaims{
 			ExpiresAt: jwt.NewNumericDate(expirationTime),
-			Issuer: "customer-order-api",
-			Subject: email,
+			Issuer:    "customer-order-api",
+			Subject:   email,
 		},
 	}
 
@@ -38,46 +38,45 @@ func generateTestToken(email string, secret []byte, expired bool) string {
 	return tokenString
 }
 
-
 func TestAuthMiddleware(t *testing.T) {
 	gin.SetMode(gin.TestMode)
 	secret := []byte("test-secret")
 
 	tests := []struct {
-		name string
-		authHeader string
+		name           string
+		authHeader     string
 		expectedStatus int
-		expectedError string
+		expectedError  string
 	}{
 		{
-			name: "missing authorization header",
-			authHeader: "",
+			name:           "missing authorization header",
+			authHeader:     "",
 			expectedStatus: http.StatusUnauthorized,
-			expectedError: "missing token",
+			expectedError:  "missing token",
 		},
 		{
-			name: "invalid authorization header format",
-			authHeader: "invalidformat token123",
+			name:           "invalid authorization header format",
+			authHeader:     "invalidformat token123",
 			expectedStatus: http.StatusUnauthorized,
-			expectedError: "invalid token format",
+			expectedError:  "invalid token format",
 		},
 		{
-			name: "missing authorization header",
-			authHeader: "",
+			name:           "missing authorization header",
+			authHeader:     "",
 			expectedStatus: http.StatusUnauthorized,
-			expectedError: "missing token",
+			expectedError:  "missing token",
 		},
 		{
-			name: "missing bearer prefix",
-			authHeader: "token123",
+			name:           "missing bearer prefix",
+			authHeader:     "token123",
 			expectedStatus: http.StatusUnauthorized,
-			expectedError: "invalid token format",
+			expectedError:  "invalid token format",
 		},
 		{
-			name: "valid token",
-			authHeader: "Bearer "+ generateTestToken("test@example.com", secret, true),
+			name:           "valid token",
+			authHeader:     "Bearer " + generateTestToken("test@example.com", secret, true),
 			expectedStatus: http.StatusUnauthorized,
-			expectedError: "missing token",
+			expectedError:  "missing token",
 		},
 		{
 			name:           "Invalid token",
@@ -138,8 +137,42 @@ func TestCORSMiddleware(t *testing.T) {
 		router.ServeHTTP(w, req)
 
 		assert.Equal(t, http.StatusOK, w.Code)
-		assert.Equal(t, "GET", w.Header().Get("Access-Control-Allow-Origin"))
-		// assert.Contains(t, w.Header().Get("Access-Control-Allow-Origin"), "GET")
-		// assert.Contains(t, w.Header().Get("Access-Control-Allow-Headers"), "Authorization")
+		assert.Equal(t, "*", w.Header().Get("Access-Control-Allow-Origin"))
 	})
+}
+
+func TestAuthMiddlewareContext(t *testing.T) {
+	gin.SetMode(gin.TestMode)
+	secret := []byte("test-secret")
+	email := "test@example.com"
+	token := generateTestToken(email, secret, false)
+
+	router := gin.New()
+	router.Use(AuthMiddleware())
+	router.GET("/test", func(ctx *gin.Context) {
+		claims, exists := ctx.Get("claims")
+		assert.True(t, exists)
+
+		userEmail, exists := ctx.Get("user_email")
+		assert.True(t, exists)
+		assert.Equal(t, email, userEmail)
+
+		userSub, exists := ctx.Get("user_sub")
+		assert.True(t, exists)
+		assert.Equal(t, email, userSub)
+
+		ctx.JSON(http.StatusOK, gin.H{
+			"message": "success",
+			"email":   claims.(*handlers.Claims).Email,
+		})
+	})
+
+	w := httptest.NewRecorder()
+	req, _ := http.NewRequest("GET", "/test", nil)
+	req.Header.Set("Authorization", "Bearer "+token)
+
+	router.ServeHTTP(w, req)
+
+	assert.Equal(t, http.StatusOK, w.Code)
+	assert.Contains(t, w.Body.String(), email)
 }
